fix: console error when typing a tag name on PDP
Ports 1.x safe portions of #4673
Resolves #4776 console error appearing when typing a tag name on the PDP
4776 came with several other unrelated commits which remove unused code
I've left those changes intact and only removed code which would cause 1.x to break
