+++ lib/api/helpers.js
@@ -1,43 +1,6 @@
-import url from "url";
 import { Meteor } from "meteor/meteor";
 import { check, Match } from "meteor/check";
 import ReactionError from "@reactioncommerce/reaction-error";
-import { Shops } from "/lib/collections";
-
-// Array of ISO Language codes for all languages that use latin based char sets
-// list is based on this matrix http://w3c.github.io/typography/gap-analysis/language-matrix.html
-// list of lang codes https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes
-export const latinLangs = ["az", "da", "de", "en", "es", "ff", "fr", "ha", "hr", "hu", "ig", "is", "it", "jv", "ku", "ms", "nl", "no", "om", "pl", "pt", "ro", "sv", "sw", "tl", "tr", "uz", "vi", "yo"]; // eslint-disable-line max-len
-
-// dynamic import of slugify/transliteration.slugify
-let slugify;
-async function lazyLoadSlugify() {
-  let mod;
-  // getting the shops base language
-  let lang;
-  if (Meteor.isServer) {
-    lang = getPrimaryShopLang();
-  } else {
-    lang = getShopLang();
-  }
-
-  // if slugify has been loaded but the language has changed
-  // to be a non latin based language then load transliteration
-  if (slugify && slugify.name === "replace" && latinLangs.indexOf(lang) === -1) {
-    mod = await import("transliteration");
-  } else if (slugify) {
-    // if slugify/transliteration is loaded and no lang change
-    return;
-  } else if (latinLangs.indexOf(lang) >= 0) {
-    // if the shops language use latin based chars load slugify else load transliterations's slugify
-    mod = await import("slugify");
-  } else {
-    mod = await import("transliteration");
-  }
-  // slugify is exported to modules.default while transliteration is exported to modules.slugify
-  slugify = mod.default || mod.slugify;
-}
-
 
 let Media;
 if (Meteor.isClient) {
@@ -51,149 +14,6 @@ if (Meteor.isClient) {
  *
  * @namespace Helpers
  */
-
-let Reaction;
-if (Meteor.isClient) {
-  Reaction = require("/client/modules/core/main").default;
-} else {
-  Reaction = require("/imports/plugins/core/core/server/Reaction").default;
-}
-
-/**
- * @name getShopSettings
- * @method
- * @memberof Helpers
- * @return {Object} returns current shop settings
- */
-export function getShopSettings() {
-  return Reaction.getShopSettings();
-}
-
-/**
- * @name getShopName
- * @method
- * @memberof Helpers
- * @return {String} returns current shop name
- */
-export function getShopName() {
-  const shopId = Reaction.getShopId();
-  let shop;
-  if (shopId) {
-    shop = Shops.findOne({
-      _id: shopId
-    });
-    return (shop && shop.name) || "";
-  }
-
-  const domain = url.parse(Reaction.absoluteUrl()).hostname;
-
-  shop = Shops.findOne({ domains: { $in: [domain] } });
-
-  return shop ? shop.name : "";
-}
-
-/**
- * @name getShopLang
- * @method
- * @memberof Helpers
- * @return {String} returns current shop language
- */
-export function getShopLang() {
-  const shopId = Reaction.getShopId();
-  const shop = Shops.findOne({ _id: shopId });
-  return shop ? shop.language : "";
-}
-
-/**
- * @name getPrimaryShopLang
- * @method
- * @memberof Helpers
- * @return {String} returns primary shop language
- */
-export function getPrimaryShopLang() {
-  const shopId = Reaction.getPrimaryShopId();
-  const shop = Shops.findOne({ _id: shopId });
-  return shop ? shop.language : "";
-}
-
-
-/**
- * @name getShopPrefix
- * @method
- * @memberof Helpers
- * @param {String} leading - Default "/", prefix, the prefix with a leading shash
- * @return {String} returns shop url prefix
- */
-export function getShopPrefix(leading = "/") {
-  const shopId = Reaction.getShopId();
-
-  if (shopId) {
-    const shop = Shops.findOne({
-      _id: shopId
-    }, {
-      fields: {
-        _id: 1,
-        name: 1,
-        shopType: 1
-      }
-    });
-
-    const settings = Reaction.getMarketplaceSettings();
-    const slug = leading + getSlug(shop.slug || getShopName().toLowerCase());
-
-    if (shop.shopType === "primary") {
-      // If naked routes is turned off, use the shop slug for our primary shop routes
-      if (settings && settings.marketplaceNakedRoutes === false) {
-        return slug;
-      }
-
-      return "";
-    }
-
-    // If this is not the primary shop, use the shop slug in routes for this shop
-    // TODO: "/shop" should be configurable in marketplace settings
-    return `/shop${slug}`;
-  }
-
-  return "";
-}
-
-/**
- * @name getAbsoluteUrl
- * @method
- * @memberof Helpers
- * @param {String} path - path to append to absolute Url, path should be prefixed with / if necessary
- * @return {String} returns absolute url (shop prefix + path)
- */
-export function getAbsoluteUrl(path) {
-  const prefix = getShopPrefix("");
-  if (prefix) {
-    const absUrl = Reaction.absoluteUrl(`${prefix}/${path}`);
-    return absUrl;
-  }
-  const absUrl = Reaction.absoluteUrl(`${path}`);
-  return absUrl;
-}
-
-/**
- * @name getSlug
- * @method
- * @memberof Helpers
- * @summary return a slugified string using "slugify" from transliteration
- * https://www.npmjs.com/package/transliteration
- * @param  {String} slugString - string to slugify
- * @return {String} slugified string
- */
-export function getSlug(slugString) {
-  let slug;
-  Promise.await(lazyLoadSlugify());
-  if (slugString) {
-    slug = slugify(slugString.toLowerCase());
-  } else {
-    slug = "";
-  }
-  return slug;
-}
 
 /**
  * @name toCamelCase
